+++ powernap/__init__.py
@@ -1,2 +1,2 @@
 """PowerNap Linux power profile manager."""
-__version__="0.9.7"
+__version__="0.9.8"